ALDS Project Report
This document provides a comprehensive analysis and implementation details of an automated laptop distribution system developed for Coventry University's Flying Start Programme. The report includes insights into the role of Power Automate in streamlining the distribution process, an in-depth analysis section that examines the need for the system based on data insights. It provides a thorough overview of the implemented flows, their functionalities, and their integration within the system. Serving as a valuable resource, the report offers insights into the technical architecture, methodologies, and outcomes of the project, making it making it a valuable read for individuals interested in technology-driven solutions for educational programs.